server/requirements.txt
paho.mqtt
influxdb_client